refactor(test): split seturl runtime tests into module files
Split the monolithic test_node_runtime_seturl.py into separate test modules organized by component: cache_manager, coordinator, execution_engine, execution_state, parser, planners, and rollback_engine. This improves test maintainability and organization by isolating test concerns into focused modules with shared fixtures in conftest.py.
